Checking for DPI/HDR/resolution changes and updating settings/swapchain as needed Added DPI awareness Updated min windows version to 10 Locking exclusive fullscreen mode to primary display only Fixed Com leak with rvalue assignments to ComObject Fixed some warnings

+++ Matchstick/src/core/core.impl/MGDFStatisticsManagerImpl.cpp
@@ -44,7 +44,8 @@ void StatisticsManager::SetRemoteEndpoint(const std::string& endpoint) {
     const auto length = StringFromGUID2(guid, buffer, 40);
     if (length != 0) {
       LOG("StatisticsManager enabled", MGDF_LOG_LOW);
-      _sessionId = Resources::ToString(buffer).substr(1, length - 3);
+      _sessionId = Resources::ToString(buffer).substr(
+          1, static_cast<size_t>(length) - 3);
       _enabled = true;
     }
   }
